Add files via upload
These codes that open a .avi file which is of a tethered E. coli cell, threshold it, fit an ellipse to the object in that file, and measure the center of mass.
 
The user needs to download all these codes plus the attached Rotating_cell.avi file into one folder. Add that folder to the MATLAB path and run readavi.m You can do that by simply writing readavi in your command window and pressing enter. This script reads the .avi file and stores variables in your workspace.
 
Next, run Tetheredcell_analysis.m:  this script will ask you to draw a crop box around an E. coli cell and then ask if you are happy with the outcome using a Boolean logic. Once you are input that you are happy it will store more variables in the workspace. You can then use the command plot(Xcenter, Ycenter) or scatter(Xcenter, Ycenter) to get the locations of the center of mass. This plot should look similar to com.fig.